% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{ssh_askpass}
\alias{ssh_askpass}
\title{SSH ASKPASS}
\usage{
ssh_askpass()
}
\description{
This returns the path to the native askpass executable which can be
used by git or ssh-agent. On Windows and MacOS the package uses this
to automatically set the \code{SSH_ASKPASS} and \code{GIT_ASKPASS} variables on
load (if not already set). Users don't have worry about this.
}
